global code refactoring
