+++ custom_components/smartir/climate.py
@@ -4,7 +4,7 @@ import logging
 import voluptuous as vol
 from numbers import Number
 
-from homeassistant.components.climate import ClimateEntity, PLATFORM_SCHEMA
+from homeassistant.components.climate import ClimateEntity
 from homeassistant.components.climate.const import (
     ClimateEntityFeature,
     HVACMode,
@@ -25,46 +25,28 @@ from homeassistant.const import (
     UnitOfTemperature,
 )
 from homeassistant.core import HomeAssistant, Event, EventStateChangedData, callback
-from homeassistant.helpers.event import async_track_state_change_event, async_call_later
+from homeassistant.helpers.event import async_track_state_change_event
 import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.typing import ConfigType
 from homeassistant.util.unit_conversion import TemperatureConverter
-from . import DeviceData
-from .controller import get_controller, get_controller_schema
+from .smartir_helpers import closest_match_value
+from .smartir_entity import load_device_data_file, SmartIR, PLATFORM_SCHEMA
 
 _LOGGER = logging.getLogger(__name__)
 
 DEFAULT_NAME = "SmartIR Climate"
-DEFAULT_DELAY = 0.5
-DEFAULT_POWER_SENSOR_DELAY = 10
 
-CONF_UNIQUE_ID = "unique_id"
-CONF_DEVICE_CODE = "device_code"
-CONF_CONTROLLER_DATA = "controller_data"
-CONF_DELAY = "delay"
 CONF_TEMPERATURE_SENSOR = "temperature_sensor"
 CONF_HUMIDITY_SENSOR = "humidity_sensor"
-CONF_POWER_SENSOR = "power_sensor"
-CONF_POWER_SENSOR_DELAY = "power_sensor_delay"
-CONF_POWER_SENSOR_RESTORE_STATE = "power_sensor_restore_state"
 
 PRECISION_DOUBLE = 2
 
 PLATFORM_SCHEMA = PLATFORM_SCHEMA.extend(
     {
-        vol.Optional(CONF_UNIQUE_ID): cv.string,
         vol.Optional(CONF_NAME, default=DEFAULT_NAME): cv.string,
-        vol.Required(CONF_DEVICE_CODE): cv.positive_int,
-        vol.Required(CONF_CONTROLLER_DATA): get_controller_schema(vol, cv),
-        vol.Optional(CONF_DELAY, default=DEFAULT_DELAY): cv.positive_float,
         vol.Optional(CONF_TEMPERATURE_SENSOR): cv.entity_id,
         vol.Optional(CONF_HUMIDITY_SENSOR): cv.entity_id,
-        vol.Optional(CONF_POWER_SENSOR): cv.entity_id,
-        vol.Optional(
-            CONF_POWER_SENSOR_DELAY, default=DEFAULT_POWER_SENSOR_DELAY
-        ): cv.positive_int,
-        vol.Optional(CONF_POWER_SENSOR_RESTORE_STATE, default=True): cv.boolean,
     }
 )
 
@@ -75,8 +57,8 @@ async def async_setup_platform(
     """Set up the IR Climate platform."""
     _LOGGER.debug("Setting up the SmartIR climate platform")
     if not (
-        device_data := await DeviceData.load_file(
-            config.get(CONF_DEVICE_CODE),
+        device_data := await load_device_data_file(
+            config,
             "climate",
             {
                 "hvac_modes": [mode for mode in HVAC_MODES if mode != HVACMode.OFF],
@@ -90,37 +72,22 @@ async def async_setup_platform(
     async_add_entities([SmartIRClimate(hass, config, device_data)])
 
 
-class SmartIRClimate(ClimateEntity, RestoreEntity):
-    _attr_should_poll = False
-    _attr_assumed_state = True
+class SmartIRClimate(SmartIR, ClimateEntity, RestoreEntity):
     _enable_turn_on_off_backwards_compatibility = False
 
-    def __init__(self, hass, config, device_data):
-        _LOGGER.debug(
-            "SmartIRClimate init started for device %s supported models %s",
-            config.get(CONF_NAME),
-            device_data["supportedModels"],
-        )
-        self.hass = hass
-        self._unique_id = config.get(CONF_UNIQUE_ID)
-        self._name = config.get(CONF_NAME)
-        self._device_code = config.get(CONF_DEVICE_CODE)
-        self._controller_data = config.get(CONF_CONTROLLER_DATA)
-        self._delay = config.get(CONF_DELAY)
+    def __init__(self, hass: HomeAssistant, config: ConfigType, device_data):
+        # Initialize SmartIR device
+        SmartIR.__init__(self, hass, config, device_data)
+
         self._temperature_sensor = config.get(CONF_TEMPERATURE_SENSOR)
         self._humidity_sensor = config.get(CONF_HUMIDITY_SENSOR)
-        self._power_sensor = config.get(CONF_POWER_SENSOR)
-        self._power_sensor_delay = config.get(CONF_POWER_SENSOR_DELAY)
-        self._power_sensor_restore_state = config.get(CONF_POWER_SENSOR_RESTORE_STATE)
         self._temperature_unit = hass.config.units.temperature_unit
 
-        self._state = STATE_OFF
         self._hvac_mode = None
         self._preset_mode = None
         self._fan_mode = None
         self._swing_mode = None
         self._hvac_action = None
-        self._on_by_remote = False
         self._current_temperature = None
         self._current_humidity = None
         self._support_flags = (
@@ -128,14 +95,6 @@ class SmartIRClimate(ClimateEntity, RestoreEntity):
             | ClimateEntityFeature.TURN_ON
             | ClimateEntityFeature.TURN_OFF
         )
-        self._power_sensor_check_expect = None
-        self._power_sensor_check_cancel = None
-
-        self._manufacturer = device_data["manufacturer"]
-        self._supported_models = device_data["supportedModels"]
-        self._supported_controller = device_data["supportedController"]
-        self._commands_encoding = device_data["commandsEncoding"]
-        self._commands = device_data["commands"]
 
         # current temperature sensor precision
         self._ha_temperature_unit = hass.config.units.temperature_unit
@@ -218,17 +177,6 @@ class SmartIRClimate(ClimateEntity, RestoreEntity):
             self._support_flags = self._support_flags | ClimateEntityFeature.SWING_MODE
             self._swing_mode = self._swing_modes[0]
 
-        # Init exclusive lock for sending IR commands
-        self._temp_lock = asyncio.Lock()
-
-        # Init the IR/RF controller
-        self._controller = get_controller(
-            self.hass,
-            self._supported_controller,
-            self._commands_encoding,
-            self._controller_data,
-        )
-
     async def async_added_to_hass(self):
         """Run when entity about to be added."""
         await super().async_added_to_hass()
@@ -239,11 +187,6 @@ class SmartIRClimate(ClimateEntity, RestoreEntity):
         last_state = await self.async_get_last_state()
 
         if last_state is not None:
-            if last_state.state == STATE_OFF:
-                self._state = STATE_OFF
-            else:
-                self._state = STATE_ON
-
             if last_state.attributes.get("hvac_mode") in self._hvac_modes:
                 self._hvac_mode = last_state.attributes.get("hvac_mode")
 
@@ -273,9 +216,6 @@ class SmartIRClimate(ClimateEntity, RestoreEntity):
             ):
                 self._target_temperature = temperature
 
-            if self._power_sensor:
-                self._on_by_remote = last_state.attributes.get("on_by_remote", False)
-
         if self._temperature_sensor:
             async_track_state_change_event(
                 self.hass, self._temperature_sensor, self._async_temp_sensor_changed
@@ -291,21 +231,6 @@ class SmartIRClimate(ClimateEntity, RestoreEntity):
             )
             if last_state is not None:
                 self._current_humidity = last_state.attributes.get("current_humidity")
-
-        if self._power_sensor:
-            async_track_state_change_event(
-                self.hass, self._power_sensor, self._async_power_sensor_changed
-            )
-
-    @property
-    def unique_id(self):
-        """Return a unique ID."""
-        return self._unique_id
-
-    @property
-    def name(self):
-        """Return the name of the climate device."""
-        return self._name
 
     @property
     def state(self):
@@ -715,21 +640,10 @@ class SmartIRClimate(ClimateEntity, RestoreEntity):
                         if "-" in commands.keys():
                             temperature = "-"
                             commands = commands["-"]
-                        elif (
-                            target_temperature is not None
-                            and (
-                                temp := sorted(
-                                    commands.keys(),
-                                    key=lambda value: abs(
-                                        float(value) - target_temperature
-                                    ),
-                                )
-                            )
-                            and len(temp)
-                        ):
+                        elif ( temp := closest_match_value(target_temperature, commands.keys()) and temp is not None ):
                             # convert selected device temperature back to HA units
                             temp_ha = convert_temp(
-                                temp[0],
+                                temp,
                                 self._data_temperature_unit,
                                 self._ha_temperature_unit,
                                 self._temp_step,
@@ -813,29 +727,6 @@ class SmartIRClimate(ClimateEntity, RestoreEntity):
         await self._async_update_hvac_action()
         self.async_write_ha_state()
 
-    async def _async_power_sensor_changed(
-        self, event: Event[EventStateChangedData]
-    ) -> None:
-        """Handle power sensor changes."""
-        old_state = event.data["old_state"]
-        new_state = event.data["new_state"]
-        if new_state is None:
-            return
-
-        if old_state is not None and new_state.state == old_state.state:
-            return
-
-        if new_state.state == STATE_ON and self._state != STATE_ON:
-            self._state = STATE_ON
-            self._on_by_remote = True
-            await self._async_update_hvac_action()
-        elif new_state.state == STATE_OFF:
-            self._on_by_remote = False
-            if self._state != STATE_OFF:
-                self._state = STATE_OFF
-                await self._async_update_hvac_action()
-        self.async_write_ha_state()
-
     async def _async_update_hvac_action(self):
         """Update thermostat current hvac action."""
         if not self._temperature_sensor:
@@ -893,45 +784,6 @@ class SmartIRClimate(ClimateEntity, RestoreEntity):
                 self._current_humidity = float(state.state)
         except ValueError as ex:
             _LOGGER.error("Unable to update from humidity sensor: %s", ex)
-
-    @callback
-    def _async_power_sensor_check_schedule(self, state):
-        if self._power_sensor_check_cancel:
-            self._power_sensor_check_cancel()
-            self._power_sensor_check_cancel = None
-            self._power_sensor_check_expect = None
-
-        @callback
-        def _async_power_sensor_check(*_):
-            self._power_sensor_check_cancel = None
-            expected_state = self._power_sensor_check_expect
-            self._power_sensor_check_expect = None
-            current_state = getattr(
-                self.hass.states.get(self._power_sensor), "state", None
-            )
-            _LOGGER.debug(
-                "Executing power sensor check for expected state '%s', current state '%s'.",
-                expected_state,
-                current_state,
-            )
-
-            if (
-                expected_state in [STATE_ON, STATE_OFF]
-                and current_state in [STATE_ON, STATE_OFF]
-                and expected_state != current_state
-            ):
-                self._state = current_state
-                _LOGGER.debug(
-                    "Power sensor check failed, reverted device state to '%s'.",
-                    self._state,
-                )
-                self.async_write_ha_state()
-
-        self._power_sensor_check_expect = state
-        self._power_sensor_check_cancel = async_call_later(
-            self.hass, self._power_sensor_delay, _async_power_sensor_check
-        )
-        _LOGGER.debug("Scheduled power sensor check for '%s' state", state)
 
 
 def convert_temp(temperature: Number, from_unit: str, to_unit: str, precision: Number):
