+++ custom_components/smartir/fan.py
@@ -6,7 +6,6 @@ import voluptuous as vol
 from homeassistant.components.fan import (
     FanEntity,
     FanEntityFeature,
-    PLATFORM_SCHEMA,
     DIRECTION_REVERSE,
     DIRECTION_FORWARD,
 )
@@ -20,37 +19,17 @@ from homeassistant.util.percentage import (
     ordered_list_item_to_percentage,
     percentage_to_ordered_list_item,
 )
-from . import DeviceData
-from .controller import get_controller, get_controller_schema
+from .smartir_entity import load_device_data_file, SmartIR, PLATFORM_SCHEMA
 
 _LOGGER = logging.getLogger(__name__)
 
 DEFAULT_NAME = "SmartIR Fan"
-DEFAULT_DELAY = 0.5
-DEFAULT_POWER_SENSOR_DELAY = 10
-
-CONF_UNIQUE_ID = "unique_id"
-CONF_DEVICE_CODE = "device_code"
-CONF_CONTROLLER_DATA = "controller_data"
-CONF_DELAY = "delay"
-CONF_POWER_SENSOR = "power_sensor"
-CONF_POWER_SENSOR_DELAY = "power_sensor_delay"
-CONF_POWER_SENSOR_RESTORE_STATE = "power_sensor_restore_state"
 
 OSCILLATING = "oscillate"
 
 PLATFORM_SCHEMA = PLATFORM_SCHEMA.extend(
     {
-        vol.Optional(CONF_UNIQUE_ID): cv.string,
         vol.Optional(CONF_NAME, default=DEFAULT_NAME): cv.string,
-        vol.Required(CONF_DEVICE_CODE): cv.positive_int,
-        vol.Required(CONF_CONTROLLER_DATA): get_controller_schema(vol, cv),
-        vol.Optional(CONF_DELAY, default=DEFAULT_DELAY): cv.positive_float,
-        vol.Optional(CONF_POWER_SENSOR): cv.entity_id,
-        vol.Optional(
-            CONF_POWER_SENSOR_DELAY, default=DEFAULT_POWER_SENSOR_DELAY
-        ): cv.positive_int,
-        vol.Optional(CONF_POWER_SENSOR_RESTORE_STATE, default=True): cv.boolean,
     }
 )
 
@@ -61,8 +40,8 @@ async def async_setup_platform(
     """Set up the IR Fan platform."""
     _LOGGER.debug("Setting up the SmartIR fan platform")
     if not (
-        device_data := await DeviceData.load_file(
-            config.get(CONF_DEVICE_CODE),
+        device_data := await load_device_data_file(
+            config,
             "fan",
             {},
             hass,
@@ -75,22 +54,12 @@ async def async_setup_platform(
 
 
 class SmartIRFan(FanEntity, RestoreEntity):
-    _attr_should_poll = False
-    _attr_assumed_state = True
     _enable_turn_on_off_backwards_compatibility = False
 
-    def __init__(self, hass, config, device_data):
-        self.hass = hass
-        self._unique_id = config.get(CONF_UNIQUE_ID)
-        self._name = config.get(CONF_NAME)
-        self._device_code = config.get(CONF_DEVICE_CODE)
-        self._controller_data = config.get(CONF_CONTROLLER_DATA)
-        self._delay = config.get(CONF_DELAY)
-        self._power_sensor = config.get(CONF_POWER_SENSOR)
-        self._power_sensor_delay = config.get(CONF_POWER_SENSOR_DELAY)
-        self._power_sensor_restore_state = config.get(CONF_POWER_SENSOR_RESTORE_STATE)
+    def __init__(self, hass: HomeAssistant, config: ConfigType, device_data):
+        # Initialize SmartIR device
+        SmartIR.__init__(self, hass, config, device_data)
 
-        self._state = STATE_OFF
         self._speed = None
         self._oscillating = None
         self._on_by_remote = False
@@ -99,14 +68,6 @@ class SmartIRFan(FanEntity, RestoreEntity):
             | FanEntityFeature.TURN_ON
             | FanEntityFeature.TURN_OFF
         )
-        self._power_sensor_check_expect = None
-        self._power_sensor_check_cancel = None
-
-        self._manufacturer = device_data["manufacturer"]
-        self._supported_models = device_data["supportedModels"]
-        self._supported_controller = device_data["supportedController"]
-        self._commands_encoding = device_data["commandsEncoding"]
-        self._commands = device_data["commands"]
 
         # fan speeds
         self._speed_list = device_data["speed"]
@@ -127,17 +88,6 @@ class SmartIRFan(FanEntity, RestoreEntity):
             self._oscillating = False
             self._support_flags = self._support_flags | FanEntityFeature.OSCILLATE
 
-        # Init exclusive lock for sending IR commands
-        self._temp_lock = asyncio.Lock()
-
-        # Init the IR/RF controller
-        self._controller = get_controller(
-            self.hass,
-            self._supported_controller,
-            self._commands_encoding,
-            self._controller_data,
-        )
-
     async def async_added_to_hass(self):
         """Run when entity about to be added."""
         await super().async_added_to_hass()
@@ -145,8 +95,6 @@ class SmartIRFan(FanEntity, RestoreEntity):
         last_state = await self.async_get_last_state()
 
         if last_state is not None:
-            self._state = last_state.state
-
             if (
                 self._support_flags & FanEntityFeature.SET_SPEED
                 and last_state.attributes.get("speed") in self._speed_list
@@ -160,29 +108,6 @@ class SmartIRFan(FanEntity, RestoreEntity):
 
             if self._support_flags & FanEntityFeature.OSCILLATE:
                 self._oscillating = last_state.attributes.get("oscillating", False)
-
-            if self._power_sensor:
-                self._on_by_remote = last_state.attributes.get("on_by_remote", False)
-
-        if self._power_sensor:
-            async_track_state_change_event(
-                self.hass, self._power_sensor, self._async_power_sensor_changed
-            )
-
-    @property
-    def unique_id(self):
-        """Return a unique ID."""
-        return self._unique_id
-
-    @property
-    def name(self):
-        """Return the display name of the fan."""
-        return self._name
-
-    @property
-    def state(self):
-        """Return the current state."""
-        return self._state
 
     @property
     def percentage(self):
@@ -214,11 +139,6 @@ class SmartIRFan(FanEntity, RestoreEntity):
             return None
         else:
             return self._current_direction
-
-    @property
-    def supported_features(self):
-        """Return the list of supported features."""
-        return self._support_flags
 
     @property
     def extra_state_attributes(self):
@@ -328,63 +248,3 @@ class SmartIRFan(FanEntity, RestoreEntity):
                 _LOGGER.exception(
                     "Exception raised in the in the _send_command '%s'", e
                 )
-
-    async def _async_power_sensor_changed(
-        self, event: Event[EventStateChangedData]
-    ) -> None:
-        """Handle power sensor changes."""
-        old_state = event.data["old_state"]
-        new_state = event.data["new_state"]
-        if new_state is None:
-            return
-
-        if old_state is not None and new_state.state == old_state.state:
-            return
-
-        if new_state.state == STATE_ON and self._state != STATE_ON:
-            self._state = STATE_ON
-            self._on_by_remote = True
-        elif new_state.state == STATE_OFF:
-            self._on_by_remote = False
-            if self._state != STATE_OFF:
-                self._state = STATE_OFF
-        self.async_write_ha_state()
-
-    @callback
-    def _async_power_sensor_check_schedule(self, state):
-        if self._power_sensor_check_cancel:
-            self._power_sensor_check_cancel()
-            self._power_sensor_check_cancel = None
-            self._power_sensor_check_expect = None
-
-        @callback
-        def _async_power_sensor_check(*_):
-            self._power_sensor_check_cancel = None
-            expected_state = self._power_sensor_check_expect
-            self._power_sensor_check_expect = None
-            current_state = getattr(
-                self.hass.states.get(self._power_sensor), "state", None
-            )
-            _LOGGER.debug(
-                "Executing power sensor check for expected state '%s', current state '%s'.",
-                expected_state,
-                current_state,
-            )
-
-            if (
-                expected_state in [STATE_ON, STATE_OFF]
-                and current_state in [STATE_ON, STATE_OFF]
-                and expected_state != current_state
-            ):
-                self._state = current_state
-                _LOGGER.debug(
-                    "Power sensor check failed, reverted device state to '%s'.",
-                    self._state,
-                )
-                self.async_write_ha_state()
-
-        self._power_sensor_check_expect = state
-        self._power_sensor_check_cancel = async_call_later(
-            self.hass, self._power_sensor_delay, _async_power_sensor_check
-        )
-        _LOGGER.debug("Scheduled power sensor check for '%s' state.", state)
