+++ custom_components/smartir/media_player.py
@@ -3,7 +3,7 @@ import logging
 
 import voluptuous as vol
 
-from homeassistant.components.media_player import MediaPlayerEntity, PLATFORM_SCHEMA
+from homeassistant.components.media_player import MediaPlayerEntity
 from homeassistant.components.media_player.const import (
     MediaPlayerEntityFeature,
     MediaType,
@@ -14,38 +14,19 @@ from homeassistant.helpers.event import async_track_state_change_event, async_ca
 import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.typing import ConfigType
-from . import DeviceData
-from .controller import get_controller, get_controller_schema
+from .smartir_entity import load_device_data_file, SmartIR, PLATFORM_SCHEMA
 
 _LOGGER = logging.getLogger(__name__)
 
 DEFAULT_NAME = "SmartIR Media Player"
 DEFAULT_DEVICE_CLASS = "tv"
-DEFAULT_DELAY = 0.5
-DEFAULT_POWER_SENSOR_DELAY = 10
 
-CONF_UNIQUE_ID = "unique_id"
-CONF_DEVICE_CODE = "device_code"
-CONF_CONTROLLER_DATA = "controller_data"
-CONF_DELAY = "delay"
-CONF_POWER_SENSOR = "power_sensor"
-CONF_POWER_SENSOR_DELAY = "power_sensor_delay"
-CONF_POWER_SENSOR_RESTORE_STATE = "power_sensor_restore_state"
 CONF_SOURCE_NAMES = "source_names"
 CONF_DEVICE_CLASS = "device_class"
 
 PLATFORM_SCHEMA = PLATFORM_SCHEMA.extend(
     {
-        vol.Optional(CONF_UNIQUE_ID): cv.string,
         vol.Optional(CONF_NAME, default=DEFAULT_NAME): cv.string,
-        vol.Required(CONF_DEVICE_CODE): cv.positive_int,
-        vol.Required(CONF_CONTROLLER_DATA): get_controller_schema(vol, cv),
-        vol.Optional(CONF_DELAY, default=DEFAULT_DELAY): cv.positive_float,
-        vol.Optional(CONF_POWER_SENSOR): cv.entity_id,
-        vol.Optional(
-            CONF_POWER_SENSOR_DELAY, default=DEFAULT_POWER_SENSOR_DELAY
-        ): cv.positive_int,
-        vol.Optional(CONF_POWER_SENSOR_RESTORE_STATE, default=True): cv.boolean,
         vol.Optional(CONF_SOURCE_NAMES): dict,
         vol.Optional(CONF_DEVICE_CLASS, default=DEFAULT_DEVICE_CLASS): cv.string,
     }
@@ -58,8 +39,8 @@ async def async_setup_platform(
     """Set up the IR Media Player platform."""
     _LOGGER.debug("Setting up the SmartIR media player platform")
     if not (
-        device_data := await DeviceData.load_file(
-            config.get(CONF_DEVICE_CODE),
+        device_data := await load_device_data_file(
+            config,
             "media_player",
             {},
             hass,
@@ -72,34 +53,16 @@ async def async_setup_platform(
 
 
 class SmartIRMediaPlayer(MediaPlayerEntity, RestoreEntity):
-    _attr_should_poll = False
-    _attr_assumed_state = True
 
-    def __init__(self, hass, config, device_data):
-        self.hass = hass
-        self._unique_id = config.get(CONF_UNIQUE_ID)
-        self._name = config.get(CONF_NAME)
-        self._device_code = config.get(CONF_DEVICE_CODE)
-        self._controller_data = config.get(CONF_CONTROLLER_DATA)
-        self._delay = config.get(CONF_DELAY)
-        self._power_sensor = config.get(CONF_POWER_SENSOR)
-        self._power_sensor_delay = config.get(CONF_POWER_SENSOR_DELAY)
-        self._power_sensor_restore_state = config.get(CONF_POWER_SENSOR_RESTORE_STATE)
+    def __init__(self, hass: HomeAssistant, config: ConfigType, device_data):
+        # Initialize SmartIR device
+        SmartIR.__init__(self, hass, config, device_data)
+
         self._device_class = config.get(CONF_DEVICE_CLASS)
 
-        self._state = STATE_OFF
         self._sources_list = []
         self._source = None
-        self._on_by_remote = False
         self._support_flags = 0
-        self._power_sensor_check_expect = None
-        self._power_sensor_check_cancel = None
-
-        self._manufacturer = device_data["manufacturer"]
-        self._supported_models = device_data["supportedModels"]
-        self._supported_controller = device_data["supportedController"]
-        self._commands_encoding = device_data["commandsEncoding"]
-        self._commands = device_data["commands"]
 
         # Supported features
         if "off" in self._commands and self._commands["off"] is not None:
@@ -158,58 +121,19 @@ class SmartIRMediaPlayer(MediaPlayerEntity, RestoreEntity):
             for key in self._commands["sources"]:
                 self._sources_list.append(key)
 
-        # Init exclusive lock for sending IR commands
-        self._temp_lock = asyncio.Lock()
-
-        # Init the IR/RF controller
-        self._controller = get_controller(
-            self.hass,
-            self._supported_controller,
-            self._commands_encoding,
-            self._controller_data,
-        )
-
     async def async_added_to_hass(self):
         """Run when entity about to be added."""
         await super().async_added_to_hass()
 
         last_state = await self.async_get_last_state()
 
-        if last_state is not None:
-            self._state = last_state.state
-
-            if self._power_sensor:
-                self._on_by_remote = last_state.attributes.get("on_by_remote", False)
-
-        if self._power_sensor:
-            async_track_state_change_event(
-                self.hass, self._power_sensor, self._async_power_sensor_changed
-            )
-
-    @property
-    def should_poll(self):
-        """Push an update after each command."""
-        return True
-
-    @property
-    def unique_id(self):
-        """Return a unique ID."""
-        return self._unique_id
-
-    @property
-    def name(self):
-        """Return the name of the media player."""
-        return self._name
+        # if last_state is not None:
+        # TODO add attributes restore
 
     @property
     def device_class(self):
         """Return the device_class of the media player."""
         return self._device_class
-
-    @property
-    def state(self):
-        """Return the state of the player."""
-        return self._state
 
     @property
     def media_title(self):
@@ -231,11 +155,6 @@ class SmartIRMediaPlayer(MediaPlayerEntity, RestoreEntity):
             return None
         else:
             return self._source
-
-    @property
-    def supported_features(self):
-        """Flag media player features that are supported."""
-        return self._support_flags
 
     @property
     def extra_state_attributes(self):
@@ -376,64 +295,3 @@ class SmartIRMediaPlayer(MediaPlayerEntity, RestoreEntity):
                 _LOGGER.exception(
                     "Exception raised in the in the _send_command '%s'", e
                 )
-
-    async def _async_power_sensor_changed(
-        self, event: Event[EventStateChangedData]
-    ) -> None:
-        """Handle power sensor changes."""
-        old_state = event.data["old_state"]
-        new_state = event.data["new_state"]
-        if new_state is None:
-            return
-
-        if old_state is not None and new_state.state == old_state.state:
-            return
-
-        if new_state.state == STATE_ON and self._state != STATE_ON:
-            self._on_by_remote = True
-            self._state = STATE_ON
-        elif new_state.state == STATE_OFF:
-            self._on_by_remote = False
-            if self._state != STATE_OFF:
-                self._state = STATE_OFF
-                # self._source = None
-        self.async_write_ha_state()
-
-    @callback
-    def _async_power_sensor_check_schedule(self, state):
-        if self._power_sensor_check_cancel:
-            self._power_sensor_check_cancel()
-            self._power_sensor_check_cancel = None
-            self._power_sensor_check_expect = None
-
-        @callback
-        def _async_power_sensor_check(*_):
-            self._power_sensor_check_cancel = None
-            expected_state = self._power_sensor_check_expect
-            self._power_sensor_check_expect = None
-            current_state = getattr(
-                self.hass.states.get(self._power_sensor), "state", None
-            )
-            _LOGGER.debug(
-                "Executing power sensor check for expected state '%s', current state '%s'.",
-                expected_state,
-                current_state,
-            )
-
-            if (
-                expected_state in [STATE_ON, STATE_OFF]
-                and current_state in [STATE_ON, STATE_OFF]
-                and expected_state != current_state
-            ):
-                self._state = current_state
-                _LOGGER.debug(
-                    "Power sensor check failed, reverted device state to '%s'.",
-                    self._state,
-                )
-                self.async_write_ha_state()
-
-        self._power_sensor_check_expect = state
-        self._power_sensor_check_cancel = async_call_later(
-            self.hass, self._power_sensor_delay, _async_power_sensor_check
-        )
-        _LOGGER.debug("Scheduled power sensor check for '%s' state.", state)
